+++ lib/helper/index.dart
@@ -1,6 +1,7 @@
 export 'player_helper.dart';
 export 'app_permission_helper.dart';
 export 'pip_helper.dart';
+export 'system_ui_helper.dart';
 export 'driver_helper.dart';
 export 'strm_helper.dart';
 export 'button_helper.dart';